started implementation of sub-controllers for clearer code.
