+++ src/main/java/GUI/LoginFrame.java
@@ -91,8 +91,9 @@ public class LoginFrame extends JFrame implements ActionListener {
                 String password = new String(passwordText.getPassword());
 
 
-                if(this.uiController.attemptLogin(id, password)){
+                if(this.uiController.loginUIControl.attemptLogin(id, password)){
                     this.dispose();
+                    this.uiController.updateUser();
                     this.uiOperator.toHome(this.uiController);
                 }
 
@@ -103,6 +104,7 @@ public class LoginFrame extends JFrame implements ActionListener {
 
         if (e.getSource() == registerBTN){
             this.dispose();
+            this.uiController.updateUser();
             this.uiOperator.toRegister(this.uiController);
         }
